Errors abundant, cause this just hates me
